update file dir
